+++ fbw-a32nx/src/systems/fmgc/src/flightplanning/FlightPlanService.ts
@@ -120,7 +120,24 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
 
     active.performanceData.pipeTo(sec.performanceData, isBeforeEngineStart);
 
+    sec.flags &= 0;
     sec.flags |= FlightPlanFlags.CopiedFromActive;
+  }
+
+  async secondaryCopyFromSecondary(from: number, to: number, isBeforeEngineStart: boolean) {
+    // TODO copy predictions
+    // We don't copy predictions for now because we don't have the computation of predictions for the secondary flight plan
+    // To keep things consistent, we don't show predictions anywhere and we don't want to show any computed turn radii
+    this.flightPlanManager.copy(
+      FlightPlanIndex.FirstSecondary + (from - 1),
+      FlightPlanIndex.FirstSecondary + (to - 1),
+      // CopyOptions.CopyPredictions,
+    );
+
+    const fromFpln = this.secondary(from);
+    const toFpln = this.secondary(to);
+
+    fromFpln.performanceData.pipeTo(toFpln.performanceData, isBeforeEngineStart);
   }
 
   async secondaryDelete(index: number) {
@@ -148,6 +165,9 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
   async activeAndSecondarySwap(secIndex: number, isBeforeEngineStart: boolean): Promise<void> {
     this.persistActivePerfDataAcrossModification(isBeforeEngineStart, () => {
       this.flightPlanManager.swap(FlightPlanIndex.FirstSecondary + secIndex - 1, FlightPlanIndex.Active);
+      const sec = this.secondary(secIndex);
+      sec.flags &= 0;
+      sec.flags |= FlightPlanFlags.SwappedWithActive;
     });
   }
 
@@ -245,6 +265,8 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       });
     }
 
+    temporaryPlan.wasModified = true;
+
     this.flightPlanManager.copy(FlightPlanIndex.Temporary, FlightPlanIndex.Active, CopyOptions.IncludeFixInfos);
     this.flightPlanManager.delete(FlightPlanIndex.Temporary);
   }
@@ -264,6 +286,8 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
 
     this.flightPlanManager.copy(FlightPlanIndex.Uplink, intoPlan);
     this.flightPlanManager.delete(FlightPlanIndex.Uplink);
+
+    this.flightPlanManager.get(intoPlan).flags |= FlightPlanFlags.CompanyFlightPlan;
 
     if (this.hasTemporary) {
       this.flightPlanManager.delete(FlightPlanIndex.Temporary);
@@ -299,6 +323,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
 
       finalIndex = FlightPlanIndex.Temporary;
     }
+    this.flightPlanManager.get(finalIndex).wasModified = true;
 
     return finalIndex;
   }
@@ -324,6 +349,8 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       await this.flightPlanManager.get(planIndex).setAlternateDestinationAirport(altnIcao);
     }
 
+    this.flightPlanManager.get(planIndex).flags |= FlightPlanFlags.ManualCreation;
+
     // Support code for TELEX API, should move somewhere else
     NXDataStore.setLegacy('PLAN_ORIGIN', fromIcao);
     NXDataStore.setLegacy('PLAN_DESTINATION', toIcao);
@@ -335,6 +362,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
     }
 
     const plan = this.flightPlanManager.get(planIndex);
+    plan.wasModified = true;
 
     if (altnIcao === undefined) {
       return plan.deleteAlternateFlightPlan();
@@ -722,6 +750,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }
 
@@ -742,6 +771,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }
 
@@ -757,6 +787,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }
 
@@ -777,6 +808,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }
 
@@ -797,6 +829,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }
 
@@ -812,6 +845,7 @@ export class FlightPlanService<P extends FlightPlanPerformanceData = FlightPlanP
       true,
     );
 
+    plan.wasModified = true;
     plan.incrementVersion();
   }