+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFpln.tsx
@@ -42,6 +42,7 @@ import { MfdFmsFplnVertRev } from 'instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFpl
 import { ConditionalComponent } from '../../../../MsfsAvionicsCommon/UiWidgets/ConditionalComponent';
 import { InternalKccuKeyEvent } from 'instruments/src/MFD/shared/MFDSimvarPublisher';
 import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
+import { ReadonlyFlightPlanElement } from '@fmgc/flightplanning/legs/ReadonlyFlightPlanLeg';
 
 interface MfdFmsFplnProps extends AbstractMfdPageProps {}
 
@@ -100,10 +101,11 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
       : '--:--',
   );
 
-  private readonly destTimeNotAvail = MappedSubject.create(
-    ([tmpy, time]) => !tmpy && time == null,
+  private readonly destTimeNotAvailAndInActive = MappedSubject.create(
+    ([tmpy, time, fpIndex]) => !tmpy && time == null && fpIndex === FlightPlanIndex.Active,
     this.tmpyActive,
     this.destTime,
+    this.loadedFlightPlanIndex,
   );
 
   private readonly destEfob = Subject.create<number | null>(null);
@@ -112,13 +114,14 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
     v !== null ? Math.max(0, Units.poundToKilogram(v) / 1_000).toFixed(1) : '---.-',
   );
 
-  private readonly destEfobNotAvailable = MappedSubject.create(
-    ([tmpy, efob]) => !tmpy && efob == null,
+  private readonly destEfobNotAvailableAndInActive = MappedSubject.create(
+    ([tmpy, efob, fpIndex]) => !tmpy && efob == null && fpIndex === FlightPlanIndex.Active,
     this.tmpyActive,
     this.destEfob,
+    this.loadedFlightPlanIndex,
   );
 
-  private readonly destEfobUnitVisiblity = this.destEfobNotAvailable.map((v) => (v ? 'hidden' : 'visible'));
+  private readonly destEfobUnitVisiblity = this.destEfobNotAvailableAndInActive.map((v) => (v ? 'hidden' : 'visible'));
 
   private readonly distanceToDest = Subject.create<number | null>(null);
 
@@ -295,7 +298,9 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
       let lastDistanceFromStart: number = 0;
       let lastLegLatLong: Coordinates = { lat: 0, long: 0 };
       // Construct leg data for all legs
-      const jointFlightPlan = this.loadedFlightPlan.allLegs.concat(this.loadedFlightPlan.alternateFlightPlan.allLegs);
+      const jointFlightPlan = this.loadedFlightPlan.allLegs.concat(
+        this.loadedAlternateFlightPlan?.allLegs as ReadonlyFlightPlanElement[],
+      );
 
       if (!jointFlightPlan.length) {
         return;
@@ -316,7 +321,10 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
 
         if (
           el instanceof FlightPlanLeg &&
-          index < (this.loadedFlightPlan?.legCount ?? 0) + (this.loadedFlightPlan?.alternateFlightPlan.legCount ?? 0)
+          index <
+            (this.loadedFlightPlan?.legCount ?? 0) +
+              (this.props.fmcService.master?.flightPlanInterface.get(this.loadedFlightPlanIndex.get())
+                .alternateFlightPlan.legCount ?? 0)
         ) {
           if (index === 0 || el.calculated === undefined) {
             newEl.distanceFromLastWpt = null;
@@ -368,7 +376,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
             isPseudoWaypoint: true,
             isAltnWaypoint: isAltn,
             isMissedAppchWaypoint: isAltn
-              ? i >= (this.loadedFlightPlan.alternateFlightPlan.firstMissedApproachLegIndex ?? Infinity)
+              ? i >= (this.loadedAlternateFlightPlan?.firstMissedApproachLegIndex ?? Infinity)
               : i >= (this.loadedFlightPlan?.firstMissedApproachLegIndex ?? Infinity),
             ident: pwp.mcduIdent ?? pwp.ident,
             overfly: false,
@@ -400,13 +408,8 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
           const transAlt = this.loadedFlightPlan.performanceData.transitionAltitude.get();
           const transLevel = this.loadedFlightPlan.performanceData.transitionLevel.get();
           const transLevelAsAlt = transLevel !== null && transLevel !== undefined ? transLevel * 100 : null;
-          const useTransLevel =
-            i >=
-            this.loadedFlightPlan.originSegment.legCount +
-              this.loadedFlightPlan.enrouteSegment.legCount +
-              this.loadedFlightPlan.departureSegment.legCount +
-              this.loadedFlightPlan.departureRunwayTransitionSegment.legCount +
-              this.loadedFlightPlan.departureEnrouteTransitionSegment.legCount;
+          const useTransLevel = i >= this.loadedFlightPlan.lastEnrouteLegIndex;
+          const firstMissedApproachLegIndex = this.loadedAlternateFlightPlan?.firstMissedApproachLegIndex ?? 0;
 
           if (leg.type === LegType.HM) {
             // Insert special HM line, TODO
@@ -416,7 +419,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
               isPseudoWaypoint: false,
               isAltnWaypoint: isAltn,
               isMissedAppchWaypoint: isAltn
-                ? i >= this.loadedFlightPlan.alternateFlightPlan.firstMissedApproachLegIndex
+                ? i >= firstMissedApproachLegIndex
                 : i >= this.loadedFlightPlan.firstMissedApproachLegIndex,
               ident: leg.definition.turnDirection === TurnDirection.Left ? 'HOLD L' : 'HOLD R',
               distFromLastWpt: leg.definition.length ?? null,
@@ -436,7 +439,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
             isPseudoWaypoint: false,
             isAltnWaypoint: isAltn,
             isMissedAppchWaypoint: isAltn
-              ? i >= this.loadedFlightPlan.alternateFlightPlan.firstMissedApproachLegIndex
+              ? i >= firstMissedApproachLegIndex
               : i >= this.loadedFlightPlan.firstMissedApproachLegIndex,
             ident: leg.ident,
             overfly: leg.definition.overfly,
@@ -484,13 +487,13 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
         if (i === this.loadedFlightPlan.allLegs.length - 1) {
           this.lineData.push(this.endOfFlightPlan);
 
-          if (this.loadedFlightPlan.alternateFlightPlan.allLegs.length === 0) {
+          if (this.loadedAlternateFlightPlan?.allLegs.length === 0) {
             this.lineData.push(this.noAlternateFlightPlan);
           }
         }
 
         // Identify end of ALTN F-PLN
-        if (this.loadedFlightPlan.alternateFlightPlan.allLegs.length > 0 && i === jointFlightPlan.length - 1) {
+        if ((this.loadedAlternateFlightPlan?.allLegs.length ?? 0) > 0 && i === jointFlightPlan.length - 1) {
           this.lineData.push({ type: FplnLineType.Special, originalLegIndex: null, label: 'END OF ALTN F-PLN' });
         }
       }
@@ -640,7 +643,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
 
   private openRevisionsMenu(legIndex: number, altnFlightPlan: boolean) {
     if (!this.revisionsMenuOpened.get()) {
-      const flightPlan = altnFlightPlan ? this.loadedFlightPlan?.alternateFlightPlan : this.loadedFlightPlan;
+      const flightPlan = altnFlightPlan ? this.loadedAlternateFlightPlan : this.loadedFlightPlan;
       const leg = flightPlan?.elementAt(legIndex);
       this.props.fmcService.master?.setRevisedWaypoint(legIndex, this.loadedFlightPlanIndex.get(), altnFlightPlan);
 
@@ -667,7 +670,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
 
   public openInsertNextWptFromWindow() {
     const flightPlan = this.props.fmcService.master?.revisedLegIsAltn.get()
-      ? this.loadedFlightPlan?.alternateFlightPlan
+      ? this.loadedAlternateFlightPlan
       : this.loadedFlightPlan;
     const wpt: NextWptInfo[] = flightPlan?.allLegs
       .map((el, idx) => {
@@ -758,8 +761,8 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
       this.destButtonLabelNode,
       this.lineColorIsTemporary,
       this.preflightPhase,
-      this.destTimeNotAvail,
-      this.destEfobNotAvailable,
+      this.destTimeNotAvailAndInActive,
+      this.destEfobNotAvailableAndInActive,
       this.distanceToDestNotAvail,
       this.distanceToDestUnitVisibility,
     );
@@ -981,7 +984,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
                   <span style="display: flex; align-items: center; justify-content: center;">*</span>
                 </div>,
               )}
-              onClick={() => this.props.fmcService.master?.flightPlanService.temporaryDelete()}
+              onClick={() => this.props.fmcService.master?.flightPlanInterface.temporaryDelete()}
               buttonStyle="color: #e68000; padding-right: 2px;"
             />
             <Button
@@ -995,7 +998,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
                   <span style="display: flex; align-items: center; justify-content: center;">*</span>
                 </div>,
               )}
-              onClick={() => this.props.fmcService.master?.flightPlanService.temporaryInsert()}
+              onClick={() => this.props.fmcService.master?.flightPlanInterface.temporaryInsert()}
               buttonStyle="color: #e68000; padding-right: 2px;"
             />
           </div>
@@ -1026,7 +1029,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
               class={{
                 'mfd-label': true,
                 'mfd-fms-yellow-text': this.lineColorIsTemporary,
-                'mfd-fms-green-text': this.destTimeNotAvail,
+                'mfd-fms-green-text': this.destTimeNotAvailAndInActive,
               }}
             >
               {this.destTimeLabel}
@@ -1036,7 +1039,7 @@ export class MfdFmsFpln extends FmsPage<MfdFmsFplnProps> {
                 class={{
                   'mfd-label': true,
                   'mfd-fms-yellow-text': this.lineColorIsTemporary,
-                  'mfd-fms-green-text': this.destEfobNotAvailable,
+                  'mfd-fms-green-text': this.destEfobNotAvailableAndInActive,
                   amber: this.destEfobAmber,
                 }}
               >